support both solo and legacy now; add legacy dataset
