+++ legacy data/Dataset75741560-818d-4bea-be5c-308c35eef62a/egos.json
@@ -1,0 +1,9 @@
+{
+  "version": "0.0.1",
+  "egos": [
+    {
+      "id": "4da92de6-d4fc-423e-a652-b5c4eb9da53a",
+      "description": ""
+    }
+  ]
+}
