+++ legacy data/Dataset75741560-818d-4bea-be5c-308c35eef62a/metric_definitions.json
@@ -1,0 +1,69 @@
+{
+  "version": "0.0.1",
+  "metric_definitions": [
+    {
+      "id": "db1b258e-d1d0-41b6-8751-16f601a2e230",
+      "name": "scenario_iteration",
+      "description": "Iteration information for dataset sequences"
+    },
+    {
+      "id": "14adb394-46c0-47e8-a3f0-99e754483b76",
+      "name": "random-seed",
+      "description": "The random seed used to initialize the random state of the simulation. Only triggered once per simulation."
+    },
+    {
+      "id": "51da3c27-369d-4929-aea6-d01614635ce2",
+      "name": "object count",
+      "description": "Counts of objects for each label in the sensor's view",
+      "spec": [
+        {
+          "label_id": 1,
+          "label_name": "Crate"
+        },
+        {
+          "label_id": 2,
+          "label_name": "Cube"
+        },
+        {
+          "label_id": 3,
+          "label_name": "Box"
+        },
+        {
+          "label_id": 4,
+          "label_name": "Terrain"
+        },
+        {
+          "label_id": 5,
+          "label_name": "Character"
+        }
+      ]
+    },
+    {
+      "id": "5ba92024-b3b7-41a7-9d3f-c03a6a8ddd01",
+      "name": "rendered object info",
+      "description": "Information about each labeled object visible to the sensor",
+      "spec": [
+        {
+          "label_id": 1,
+          "label_name": "Crate"
+        },
+        {
+          "label_id": 2,
+          "label_name": "Cube"
+        },
+        {
+          "label_id": 3,
+          "label_name": "Box"
+        },
+        {
+          "label_id": 4,
+          "label_name": "Terrain"
+        },
+        {
+          "label_id": 5,
+          "label_name": "Character"
+        }
+      ]
+    }
+  ]
+}
