+++ legacy data/Logs/heartbeat.txt
@@ -1,0 +1,1 @@
+0.050
