+++ unity_cv_datasetvisualizer/datasetvisualizer/SoloPreview.py
@@ -201,135 +201,80 @@ def display_number_frames(num_frames: int):
     st.sidebar.markdown("### Number of frames: " + str(num_frames))
 
 
-def preview_dataset(base_dataset_dir: str):
-    """
-    Adds streamlit components to the app to construct the dataset preview.
+def preview_dataset(data_root, folder_name):
+    instances = datamaker_dataset(data_root)
 
-    :param base_dataset_dir: The directory that contains the perception dataset.
-    :type base_dataset_dir: str
-    """
+    # if it is not a datamaker dataset
+    if instances is None:
+        # Attempt to read as a normal perception dataset
+        ds = Dataset(data_root)
+        if not ds.dataset_valid:
+            st.warning("The provided Dataset folder \"" + data_root + "\" is not considered valid")
 
-    # Create state with default values
-    create_session_state_data({
-        'zoom_image': '-1',
-        'start_at': '0',
-        'num_cols': '3',
-        'curr_dir': base_dataset_dir,
+            st.markdown("# Please open a dataset folder:")
+            if st.button("Open Dataset", key="second open dataset"):
+                folder_select()
+            return
 
-        'just_opened_zoom': True,
-        'just_opened_grid': True,
+        if len(folder_name) >= 1:
+            st.sidebar.markdown("# Current dataset:")
+            st.sidebar.write(folder_name)
 
-        'bbox2d_existed_last_time': False,
-        'bbox3d_existed_last_time': False,
-        'keypoints_existed_last_time': False,
-        'semantic_existed_last_time': False,
+        # SB HERE
+        dataset_len = ds.metadata["totalFrames"]
+        display_number_frames(dataset_len)
 
-        'previous_labelers': {},
-        'labelers_changed': False,
-    })
+        available_labelers = ds.get_available_labelers()
+        annotator_dic = ds.get_annotator_dictionary()
+        labelers = create_sidebar_labeler_menu(available_labelers, annotator_dic)
 
-    # Gets the latest selected directory
-    base_dataset_dir = st.session_state.curr_dir
+        # zoom_image is negative if the application isn't in zoom mode
+        index = int(st.session_state.zoom_image)
+        if index >= 0:
+            zoom(index, 0, ds, labelers, annotator_dic)
+        else:
+            num_rows = 5
+            grid_view(num_rows, ds, labelers, annotator_dic)
 
-    # Display select dataset menu
-    st.sidebar.markdown("# Select Dataset")
-    if st.sidebar.button("Open Dataset"):
-        folder_select()
-
-    if base_dataset_dir is None:
-        st.markdown("# Please open a dataset folder:")
-        if st.button("Open Dataset", key="second open dataset"):
-            folder_select()
-        return
-
-    # Display name of dataset (Name of folder)
-    dataset_name = os.path.abspath(base_dataset_dir).replace("\\", "/")
-
-    if dataset_name[-1] == '/':
-        folder_name = dataset_name.split('/')[-2]
+    # if it is a datamaker dataset
     else:
-        folder_name = dataset_name.split('/')[-1]
+        if len(folder_name) >= 1:
+            st.sidebar.markdown("# Current dataset:")
+            st.sidebar.write(folder_name)
 
-    if dataset_name is not None and dataset_name.strip() != "":
-        data_root = os.path.abspath(dataset_name)
-        # Attempt to read data_root as a datamaker dataset
-        instances = datamaker_dataset(data_root)
+        display_number_frames(datamaker.get_dataset_length_with_instances(instances))
 
-        # if it is not a datamaker dataset
-        if instances is None:
-            # Attempt to read as a normal perception dataset
-            ds = Dataset(data_root)
-            if not ds.dataset_valid:
-                st.warning("The provided Dataset folder \"" + data_root + "\" is not considered valid")
+        # zoom_image is negative if the application isn't in zoom mode
+        index = int(st.session_state.zoom_image)
+        if index >= 0:
+            instance_key = datamaker.get_instance_by_capture_idx(instances, index)
 
-                st.markdown("# Please open a dataset folder:")
-                if st.button("Open Dataset", key="second open dataset"):
-                    folder_select()
-                return
+            if (instance_key is None):
+                index = 0
+                instance_key = datamaker.get_instance_by_capture_idx(instances, index)
 
-            if len(folder_name) >= 1:
-                st.sidebar.markdown("# Current dataset:")
-                st.sidebar.write(folder_name)
-
-            # SB HERE
-            dataset_len = ds.metadata["totalFrames"]
-            display_number_frames(dataset_len)
-
-            available_labelers = ds.get_available_labelers()
+            offset = datamaker.get_dataset_length_with_instances(instances, instance_key)
+            ds = instances[instance_key]
+            ann_def = ds.ann_def
+            available_labelers = [a["name"] for a in ann_def.table.to_dict('records')]
             annotator_dic = ds.get_annotator_dictionary()
             labelers = create_sidebar_labeler_menu(available_labelers, annotator_dic)
-
-            # zoom_image is negative if the application isn't in zoom mode
-            index = int(st.session_state.zoom_image)
-            if index >= 0:
-                zoom(index, 0, ds, labelers, annotator_dic)
-            else:
-                num_rows = 5
-                grid_view(num_rows, ds, labelers, annotator_dic)
-
-        # if it is a datamaker dataset
+            zoom(index, offset, ds, labelers, annotator_dic)
         else:
-            if len(folder_name) >= 1:
-                st.sidebar.markdown("# Current dataset:")
-                st.sidebar.write(folder_name)
+            index = st.session_state.start_at
+            num_rows = 5
+            instance_key = datamaker.get_instance_by_capture_idx(instances, index)
 
-            display_number_frames(datamaker.get_dataset_length_with_instances(instances))
-
-            # zoom_image is negative if the application isn't in zoom mode
-            index = int(st.session_state.zoom_image)
-            if index >= 0:
+            if (instance_key is None):
+                st.session_state.start_at = 0
+                index = 0
                 instance_key = datamaker.get_instance_by_capture_idx(instances, index)
 
-                if (instance_key is None):
-                    index = 0
-                    instance_key = datamaker.get_instance_by_capture_idx(instances, index)
-
-                offset = datamaker.get_dataset_length_with_instances(instances, instance_key)
-                ds = instances[instance_key]
-                ann_def = ds.ann_def
-                available_labelers = [a["name"] for a in ann_def.table.to_dict('records')]
-                annotator_dic = ds.get_annotator_dictionary()
-                labelers = create_sidebar_labeler_menu(available_labelers, annotator_dic)
-                zoom(index, offset, ds, labelers, annotator_dic)
-            else:
-                index = st.session_state.start_at
-                num_rows = 5
-                instance_key = datamaker.get_instance_by_capture_idx(instances, index)
-
-                if (instance_key is None):
-                    st.session_state.start_at = 0
-                    index = 0
-                    instance_key = datamaker.get_instance_by_capture_idx(instances, index)
-
-                ds = instances[instance_key]
-                ann_def = ds.ann_def
-                available_labelers = [a["name"] for a in ann_def.table.to_dict('records')]
-                labelers = create_sidebar_labeler_menu(available_labelers)
-                grid_view_instances(num_rows, instances, labelers)
-    else:
-        st.markdown("# Please select a valid dataset folder:")
-        if st.button("Select dataset folder"):
-            folder_select()
+            ds = instances[instance_key]
+            ann_def = ds.ann_def
+            available_labelers = [a["name"] for a in ann_def.table.to_dict('records')]
+            labelers = create_sidebar_labeler_menu(available_labelers)
+            grid_view_instances(num_rows, instances, labelers)
 
 
 def folder_select():
@@ -543,30 +488,3 @@ def zoom(index: int,
         for i in labeler_annotations:
             st.markdown("## " + i['id'])
             st.write(i)
-
-
-def preview_app(args):
-    """
-    Starts the dataset preview app.
-
-    :param args: Arguments for the app, such as dataset
-    :type args: Namespace
-    """
-    preview_dataset(args["data"])
-
-
-if __name__ == "__main__":
-
-    # This needs to be the first streamlit command
-    st.set_page_config(layout="wide")
-    # removes the default zoom button on images
-    st.markdown('<style>button.css-enefr8{display: none;}'
-                '       button.css-1u96g9d{display: none;}</style>', unsafe_allow_html=True)
-
-    parser = argparse.ArgumentParser()
-    parser.add_argument("data", type=str)
-    args = parser.parse_args()
-    if os.path.isdir(args.data):
-        preview_app({"data": args.data})
-    else:
-        preview_app({"data": None})
